// src/pages/editprofile/UserProfile.type.ts
export type UserProfile = {
  id: string;
  email?: string;
  phone?: string;
  address?: string;
  type?: string;
  city?: string;
  state?: string;
  pincode?: string;
  country?: string;
  profile_pic?: string;
  account_status?: string;
  role?: string;
  dob?: string;
  username?: string;
  password?: string;
  social_media_links?: string;
  addition_details?: string;
  language?: string[];
  departMent?: string;
  designation?: string;
  time_zone?: string;
  last_login?: string;
  last_device?: string;
  last_ip?: string;
  currency?: string;
  domain_user_id?: string;
  domain_name?: string;
  plan_type?: string;
  added_by_user_id?: string;
  created_at?: string;
  updated_at?: string;
};